Problems

Thunderbolt.NullProblem — Type

NullProblem(ndofs)

Utility type to describe that Jacobian and residual are zero, but ndofs dofs are present. This is a trick for Lagrange-type couplings to describe the Lagrange dofs.

Thunderbolt.CoupledProblem — Type

CoupledProblem{MT, CT}

Generic description of a coupled problem.

Missing docstring.

Missing docstring for Thunderbolt.SplitProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.PartitionedProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.ODEProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.AbstractPointwiseProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.PointwiseODEProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.TransientHeatProblem. Check Documenter's build log for details.

Thunderbolt.QuasiStaticNonlinearProblem — Type

QuasiStaticNonlinearProblem{M <: QuasiStaticModel, DH <: Ferrite.AbstractDofHandler}

A discrete problem with time dependent terms and no time derivatives w.r.t. any solution variable. Abstractly written we want to solve the problem F(u, t) = 0 on some time interval [t₁, t₂].

Thunderbolt.QuasiStaticODEProblem — Type

QuasiStaticODEProblem{M <: QuasiStaticModel, DH <: Ferrite.AbstractDofHandler}

A problem with time dependent terms and time derivatives only w.r.t. internal solution variable.

TODO implement.

Thunderbolt.QuasiStaticDAEProblem — Type

QuasiStaticDAEProblem{M <: QuasiStaticModel, DH <: Ferrite.AbstractDofHandler}

A problem with time dependent terms and time derivatives only w.r.t. internal solution variable which can't be expressed as an ODE.

TODO implement.