Element Interface

Entry Points

FerriteOperators.AbstractVolumetricElementCache — Type

Supertype for all caches to integrate over volumes.

General Interface:

setup_element_cache(integrator, sdh)

Specialized Interface for Condensed Problems:

get_number_of_internal_dofs_per_element(model, element_cache, sdh)

FerriteOperators.AbstractSurfaceElementCache — Type

Supertype for all caches to integrate over surfaces.

Interface:

setup_boundary_cache(integrator, sdh)

Missing docstring.

Missing docstring for Thunderbolt.AbstractInterfaceElementCache. Check Documenter's build log for details.

FerriteOperators.assemble_element! — Function

assemble_element!(Kₑ::AbstractMatrix, cell::CellCache, element_cache::AbstractVolumetricElementCache, time)

Main entry point for bilinear operators

assemble_element!(Kₑ::AbstractMatrix, uₑ::AbstractVector, cell::CellCache, element_cache::AbstractVolumetricElementCache, time)

Update element matrix in nonlinear operators

assemble_element!(Kₑ::AbstractMatrix, residualₑ::AbstractVector, uₑ::AbstractVector, cell::CellCache, element_cache::AbstractVolumetricElementCache, time)

Update element matrix and residual in nonlinear operators

assemble_element!(residualₑ::AbstractVector, uₑ::AbstractVector, cell::CellCache, element_cache::AbstractVolumetricElementCache, time)

Update residual in nonlinear operators

The notation is as follows.

\[K_e\]
the element stiffness matrix

\[u_e\]
the element unknowns

\[residual_e\]
the element residual

FerriteOperators.assemble_facet! — Function

assemble_facet!(Kₑ::AbstractMatrix, cell::CellCache, face_cache::AbstractSurfaceElementCache, time)

Main entry point for bilinear operators

assemble_facet!(Kₑ::AbstractMatrix, uₑ::AbstractVector, cell::CellCache, face_cache::AbstractSurfaceElementCache, time)

Update face matrix in nonlinear operators

assemble_facet!(Kₑ::AbstractMatrix, residualₑ::AbstractVector, uₑ::AbstractVector, cell::CellCache, face_cache::AbstractSurfaceElementCache, time)

Update face matrix and residual in nonlinear operators

assemble_facet!(residualₑ::AbstractVector, uₑ::AbstractVector, cell::CellCache, face_cache::AbstractSurfaceElementCache, time)

Update residual in nonlinear operators

The notation is as follows.

\[K_e\]
the element stiffness matrix

\[u_e\]
the element unknowns

\[residual_e\]
the element residual

Missing docstring.

Missing docstring for Thunderbolt.assemble_interface!. Check Documenter's build log for details.

Common

Thunderbolt.AnalyticalCoefficientElementCache — Type

AnalyticalCoefficientElementCache(f(x,t)->..., nonzero_in_time_intervals, cellvalues)

Analytical coefficient described by a function in space and time. Can be sparse in time.

Thunderbolt.SimpleFacetCache — Type

Standard cache for surface integrals.

Composite

Missing docstring.

Missing docstring for Thunderbolt.CompositeVolumetricElementCache. Check Documenter's build log for details.

FerriteOperators.CompositeSurfaceElementCache — Type

This cache allows to combine multiple elements over the same surface.

Missing docstring.

Missing docstring for Thunderbolt.CompositeInterfaceElementCache. Check Documenter's build log for details.

Bilinear

Thunderbolt.BilinearMassIntegrator — Type

BilinearMassIntegrator{MT, CV}

Represents the integrand of the bilinearform $a(u,v) = \int \rho(x) v(x) u(x) dx$ for $u,v$ from the same function space with some given density field $\rho(x)$.

Thunderbolt.BilinearMassElementCache — Type

The cache associated with BilinearMassIntegrator to assemble element mass matrices.

Thunderbolt.BilinearDiffusionIntegrator — Type

BilinearDiffusionIntegrator{CoefficientType}

Represents the integrand of the bilinear form $a(u,v) = -\int \nabla v(x) \cdot D(x) \nabla u(x) dx$ for a given diffusion tensor $D(x)$ and $u,v$ from the same function space.

Thunderbolt.BilinearDiffusionElementCache — Type

The cache associated with BilinearDiffusionIntegrator to assemble element diffusion matrices.

Solid Mechanics

Elements

Thunderbolt.QuasiStaticElementCache — Type

QuasiStaticElementCache

A generic cache to assemble elements coming from a StructuralModel.

Right now the model has to be formulated in the first Piola Kirchhoff stress tensor and F.