Element Interface

Thunderbolt.AnalyticalCoefficientElementCache — Type

AnalyticalCoefficientElementCache(f(x,t)->..., nonzero_in_time_intervals, cellvalues)

Analytical coefficient described by a function in space and time. Can be sparse in time.

Thunderbolt.SimpleFacetCache — Type

Standard cache for surface integrals.

Bilinear

Thunderbolt.BilinearMassIntegrator — Type

BilinearMassIntegrator{MT, CV}

Represents the integrand of the bilinearform $a(u,v) = \int \rho(x) v(x) u(x) dx$ for $u,v$ from the same function space with some given density field $\rho(x)$.

Thunderbolt.BilinearMassElementCache — Type

The cache associated with BilinearMassIntegrator to assemble element mass matrices.

Thunderbolt.BilinearDiffusionIntegrator — Type

BilinearDiffusionIntegrator{CoefficientType}

Represents the integrand of the bilinear form $a(u,v) = -\int \nabla v(x) \cdot D(x) \nabla u(x) dx$ for a given diffusion tensor $D(x)$ and $u,v$ from the same function space.

Thunderbolt.BilinearDiffusionElementCache — Type

The cache associated with BilinearDiffusionIntegrator to assemble element diffusion matrices.

Solid Mechanics

Elements

Thunderbolt.QuasiStaticElementCache — Type

QuasiStaticElementCache

A generic cache to assemble elements coming from a StructuralModel.

Right now the model has to be formulated in the first Piola Kirchhoff stress tensor and F.