Functions

Thunderbolt.AbstractSemidiscreteFunction — Type

AbstractSemidiscreteFunction <: SciMLBase.AbstractDiffEqFunction{iip=true}

Supertype for all functions coming from PDE discretizations.

Interface

solution_size(::AbstractSemidiscreteFunction)
get_strategy(::AbstractSemidiscreteFunction)

Thunderbolt.AbstractSemidiscreteBlockedFunction — Type

AbstractSemidiscreteBlockedFunction <: AbstractSemidiscreteFunction

Supertype for all functions coming from PDE discretizations with blocked structure.

Interface

BlockArrays.blocksizes(::AbstractSemidiscreteFunction)
BlockArrays.blocks(::AbstractSemidiscreteFunction) -> Iterable

Missing docstring.

Missing docstring for Thunderbolt.AbstractSemidiscreteODEFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.AbstractSemidiscreteDAEFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.AbstractQuasiStaticFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TransientDiffusionFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SteadyDiffusionFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PointwiseODEFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for QuasiStaticNonlinearFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RSAFDQ20223DFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractOperatorSplitFunction. Check Documenter's build log for details.

OrdinaryDiffEqOperatorSplitting.GenericSplitFunction — Type

GenericSplitFunction(functions::Tuple, solution_indices::Tuple)
GenericSplitFunction(functions::Tuple, solution_indices::Tuple, syncronizers::Tuple)

This type of function describes a set of connected inner functions in mass-matrix form, as usually found in operator splitting procedures.