Functions

Thunderbolt.AbstractSemidiscreteFunction — Type

AbstractSemidiscreteFunction <: DiffEqBase.AbstractDiffEqFunction{iip=true}

Supertype for all functions coming from PDE discretizations.

Interface

solution_size(::AbstractSemidiscreteFunction)

Thunderbolt.AbstractSemidiscreteBlockedFunction — Type

AbstractSemidiscreteBlockedFunction <: AbstractSemidiscreteFunction

Supertype for all functions coming from PDE discretizations with blocked structure.

Interface

BlockArrays.blocksizes(::AbstractSemidiscreteFunction)
BlockArrays.blocks(::AbstractSemidiscreteFunction) -> Iterable

Missing docstring.

Missing docstring for Thunderbolt.AbstractSemidiscreteODEFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.AbstractSemidiscreteDAEFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.AbstractQuasiStaticFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TransientHeatFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PointwiseODEFunction. Check Documenter's build log for details.

Thunderbolt.QuasiStaticNonlinearFunction — Type

QuasiStaticNonlinearFunction{M <: QuasiStaticModel, DH <: Ferrite.AbstractDofHandler}

A discrete problem with time dependent terms and no time derivatives w.r.t. any solution variable. Abstractly written we want to solve the problem F(u, t) = 0 on some time interval [t₁, t₂].

Missing docstring.

Missing docstring for RSAFDQ20223DFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractOperatorSplitFunction. Check Documenter's build log for details.

Thunderbolt.OS.GenericSplitFunction — Type

GenericSplitFunction(functions::Tuple, dof_ranges::Tuple)
GenericSplitFunction(functions::Tuple, dof_ranges::Tuple, syncronizers::Tuple)

This type of function describes a set of connected inner functions in mass-matrix form, as usually found in operator splitting procedures.

Automatic sync

We should be able to get rid of the synchronizer and handle the connection of coefficients and solutions in semidiscretize.