Mesh

Thunderbolt.SimpleMesh2D — Type

SimpleMesh2D{C <: AbstractCell, T <: Real} <: AbstractGrid{2}

A grid which also has information abouts its vertices, faces and edges.

Thunderbolt.SimpleMesh3D — Type

SimpleMesh3D{C <: AbstractCell, T <: Real} <: AbstractGrid{3}

A grid which also has information abouts its vertices, faces and edges.

Missing docstring.

Missing docstring for Thunderbolt.to_mesh. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.elementtypes. Check Documenter's build log for details.

Coordinate Systems

Thunderbolt.CartesianCoordinateSystem — Type

CartesianCoordinateSystem{sdim, IP <: VectorizedInterpolation{sdim}}

Standard cartesian coordinate system.

Thunderbolt.LVCoordinateSystem — Type

LVCoordinateSystem(dh, u_transmural, u_apicobasal)

Simplified universal ventricular coordinate on LV only, containing the transmural, apicobasal and circumferential coordinates. See compute_LV_coordinate_system to construct it.

Mesh Generators

Missing docstring.

Missing docstring for generate_mesh. Check Documenter's build log for details.

Thunderbolt.generate_ring_mesh — Function

generate_ring_mesh(num_elements_circumferential::Int, num_elements_radial::Int, num_elements_logintudinal::Int; inner_radius::T = Float64(0.75), outer_radius::T = Float64(1.0), longitudinal_lower::T = Float64(-0.2), longitudinal_upper::T = Float64(0.2), apicobasal_tilt::T=Float64(0.0)) where {T}

Generates an idealized full-hexahedral ring with linear ansatz. Geometrically it is the substraction of a small cylinder $C_i$ of a large cylinder $C_o$. The number of elements for the cylindrical system can be controlled by the first three input parameters. The remaining parameters control the spatial dimensions and the ring shape.

Thunderbolt.generate_quadratic_ring_mesh — Function

generate_quadratic_ring_mesh(num_elements_circumferential::Int, num_elements_radial::Int, num_elements_logintudinal::Int; inner_radius::T = Float64(0.75), outer_radius::T = Float64(1.0), longitudinal_lower::T = Float64(-0.2), longitudinal_upper::T = Float64(0.2), apicobasal_tilt::T=Float64(0.0)) where {T}

Generates an idealized full-hexahedral ring with quadratic ansatz. Geometrically it is the substraction of a small cylinder $C_i$ of a large cylinder $C_o$. The number of elements for the cylindrical system can be controlled by the first three input parameters. The remaining parameters control the spatial dimensions and the ring shape.

Thunderbolt.generate_ideal_lv_mesh — Function

generate_ideal_lv_mesh(num_elements_circumferential::Int, num_elements_radial::Int, num_elements_logintudinally::Int; inner_chamber_radius = 0.7, outer_wall_radius = 1.0, longitudinal_cutoff_lower::T = Float64(-1.0), longitudinal_cutoff_upper::T = Float64(0.2), longitudinal_stretch::T = Float64(1.2))

Generate an idealized left ventricle as a truncated ellipsoid. The number of elements per axis are controlled by the first three parameters.

Utility

Missing docstring.

Missing docstring for Thunderbolt.hexahedralize. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.uniform_refinement. Check Documenter's build log for details.

Thunderbolt.load_carp_mesh — Function

load_carp_mesh(filename)

Mesh format taken from https://carp.medunigraz.at/file_formats.html .

Thunderbolt.load_voom2_mesh — Function

load_voom2_mesh(filename)

Loader for the voom2 legacy format.

Thunderbolt.load_mfem_mesh — Function

load_mfem_mesh(filename)

Loader for straight mfem meshes supporting v1.0.