Solver

Linear

Thunderbolt.SchurComplementLinearSolver — Type

SchurComplementLinearSolver(inner_alg::AbstractLinearAlgorithm)

A solver for block systems of the form

\[\begin{bmatrix} A_{11} & A_{12} \\ A_{21} & 0 \end{bmatrix} \begin{bmatrix} u_1 \\ u_2 \end{bmatrix} = \begin{bmatrix} b_1 \\ b_2 \end{bmatrix}\]

with small zero block of size $N_2 \times N_2$ and invertible $A_{11}$ with size $N_1 \times N_1$. The inner linear solver is responsible to solve for $N_2$ systems of the form $A_{11} z_i = c_i$.

Nonlinear

Thunderbolt.NewtonRaphsonSolver — Type

NewtonRaphsonSolver{T}

Classical Newton-Raphson solver to solve nonlinear problems of the form F(u) = 0. To use the Newton-Raphson solver you have to dispatch on

update_linearization!

Time

Missing docstring.

Missing docstring for BackwardEulerSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ForwardEulerSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ForwardEulerCellSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.AdaptiveForwardEulerReactionSubCellSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Thunderbolt.ThreadedForwardEulerCellSolver. Check Documenter's build log for details.

Thunderbolt.LoadDrivenSolver — Type

LoadDrivenSolver{IS, T, PFUN}

Solve the nonlinear problem F(u,t)=0 with given time increments Δton some interval [t_begin, t_end] where t is some pseudo-time parameter.

Operator Splitting Module

Thunderbolt.OS.LieTrotterGodunov — Type

LieTrotterGodunov <: AbstractOperatorSplittingAlgorithm

A first order operator splitting algorithm attributed to [1–3].

Missing docstring.

Missing docstring for Thunderbolt.OS.GenericSplitFunction. Check Documenter's build log for details.

Thunderbolt.OS.OperatorSplittingIntegrator — Type

OperatorSplittingIntegrator <: AbstractODEIntegrator

A variant of ODEIntegrator to perform opeartor splitting.

Derived from https://github.com/CliMA/ClimaTimeSteppers.jl/blob/ef3023747606d2750e674d321413f80638136632/src/integrators.jl.